+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## MapLibre GL widget v0.x.0
 
-- Add feature for building marker with Font Awesome icon
+- Add arc layer, scatterplot layer, icon layer, line layer and geojson layer
+  feature
+- Add hexagon layer feature (#23)
+- Add feature for building marker with Font Awesome icon (#22)
 - Add Raster layer feature (#21)
 - Improve build process (#20)